AFD2d_mod.jl Documentation

Finite-difference modeling for a constant density acoustic wave equation

\[\frac{1}{c^2(\mathbf{x})} \frac{\partial P}{\partial t} = \nabla^2 P - \rho \, \dot{q}(t) \, \delta(\mathbf{x} - \mathbf{x}_{s})\]

where $c(\mathbf{x})$ is the medium propagation velocity, $\rho$ its density, and $P(\mathbf{x},t)$ the acoustic wavefield excited by the volume injection rate $q(t)$ at the source position $\mathbf{x}_{s}$.

Authors: Jesse Costa and Joerg Schleicher

AFD2d_mod.afd_mod — Method

afd_mod(FD::AFD2d,
        T::Float64,
        x1s::Array{Float64,1},
        x2s::Array{Float64,1};
        x1r::Array{Float64,1},
        x2r::Array{Float64,1},
        section::Array{SU_trc,1}
        )

Acoustic wavefield forward modeling

Signature:

FD : finite-Difference type

T : simulation time

x1s: source stations x1-coordinates

x2s: source stations x2-coordinates

x1r: receiver stations x1-coordinates

x2r: receiver stations x2-coordinates

AFD2d_mod.Model2d — Method

Model2d(file::String)

Type Model2d stores the 2D velocity model and
extends its grid boundaries for FD computations

Constructor signature:

file: velocity model header file

AFD2d_mod.Grid2d — Type

Grid2d(n1::Int64,
       n2::Int64,
       o1::Float64,
       o2::Float64,
       d1::Float64
       d2::Float64)

Type to define a regular grid for Finite-Difference

Constructor Signature:

n1: number of grid point along dimension 1

n2: number of grid point along dimension 2

o1: x1-coordinate of grid origin

o2: x2-coordinate of grid origin

d1: sampling interval along x1-coordinate

d2: sampling interval along x2-coordinate

AFD2d_mod.SU_mod.SUGetTrace — Method

SUPutTrace(io::IOStream,trc::SU_trc)

Read a SU trace from a file

Signature

io : file stream reference

trc : seismic trace (header + data)

AFD2d_mod.SU_mod.SUPutTrace — Method

SUPutTrace(io::IOStream,trc::SU_trc)

Write a SU trace to a file

Signature

io : file stream reference

trc : seismic trace (header + data)